chore: clarify build/framework systems
